# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AuthErrorView": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "AuthErrorView",
        "type": "object"
      },
...
      "BasicSettingsView": {
        "additionalProperties": false,
        "properties": {
          "default_version": {
            "title": "Default Version",
            "type": "integer"
          },
          "descriptors": {
            "items": {
              "$ref": "#/components/schemas/SettingDescriptor"
            },
            "title": "Descriptors",
            "type": "array"
          },
          "generation_enabled": {
            "title": "Generation Enabled",
            "type": "boolean"
          },
          "policy": {
            "$ref": "#/components/schemas/PlanPolicy"
          },
          "policy_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Hash"
          },
          "revision": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision"
          },
          "revision_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision Id"
          }
        },
        "required": [
          "default_version",
          "revision_id",
          "revision",
          "policy_hash",
          "policy",
          "generation_enabled",
          "descriptors"
        ],
        "title": "BasicSettingsView",
        "type": "object"
      },
...
      "DiffItem": {
        "additionalProperties": false,
        "properties": {
          "after": {
            "title": "After"
          },
          "before": {
            "title": "Before"
          },
          "key": {
            "title": "Key",
            "type": "string"
          }
        },
        "required": [
          "key",
          "before",
          "after"
        ],
        "title": "DiffItem",
        "type": "object"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
...
      "HistoryItem": {
        "additionalProperties": false,
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "created_at": {
            "title": "Created At",
            "type": "integer"
          },
          "policy": {
            "$ref": "#/components/schemas/PlanPolicy"
          },
          "policy_hash": {
            "title": "Policy Hash",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "revision_id": {
            "format": "uuid",
            "title": "Revision Id",
            "type": "string"
          }
        },
        "required": [
          "revision_id",
          "revision",
          "policy_hash",
          "created_at",
          "active",
          "policy"
        ],
        "title": "HistoryItem",
        "type": "object"
      },
      "HistoryView": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/HistoryItem"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "HistoryView",
        "type": "object"
      },
      "ImageSize": {
        "additionalProperties": false,
        "properties": {
          "height": {
            "maximum": 8192.0,
            "minimum": 1.0,
            "title": "Height",
            "type": "integer"
          },
          "width": {
            "maximum": 8192.0,
            "minimum": 1.0,
            "title": "Width",
            "type": "integer"
          }
        },
        "required": [
          "width",
          "height"
        ],
        "title": "ImageSize",
        "type": "object"
      },
...
      "PlanPolicy": {
        "additionalProperties": false,
        "properties": {
          "active_jobs": {
            "default": 0,
            "maximum": 256.0,
            "minimum": 0.0,
            "title": "Active Jobs",
            "type": "integer"
          },
          "can_publish": {
            "default": false,
            "title": "Can Publish",
            "type": "boolean"
          },
          "capability_ids": {
            "default": [],
            "items": {
              "pattern": "^[a-z0-9][a-z0-9._-]{0,79}$",
              "type": "string"
            },
            "maxItems": 256,
            "title": "Capability Ids",
            "type": "array"
          },
          "executors": {
            "default": [],
            "items": {
              "enum": [
                "api",
                "local"
              ],
              "type": "string"
            },
            "maxItems": 2,
            "title": "Executors",
            "type": "array"
          },
          "image_sizes": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ImageSize"
            },
            "maxItems": 32,
            "title": "Image Sizes",
            "type": "array"
          },
          "input_count": {
            "default": 0,
            "maximum": 32.0,
            "minimum": 0.0,
            "title": "Input Count",
            "type": "integer"
          },
          "max_action_credits": {
            "default": 0,
            "maximum": 1000000000.0,
            "minimum": 0.0,
            "title": "Max Action Credits",
            "type": "integer"
          },
          "storage_bytes": {
            "default": 0,
            "maximum": 1000000000000.0,
            "minimum": 0.0,
            "title": "Storage Bytes",
            "type": "integer"
          },
          "submissions": {
            "default": 0,
            "maximum": 1000000.0,
            "minimum": 0.0,
            "title": "Submissions",
            "type": "integer"
          },
          "upload_bytes": {
            "default": 0,
            "maximum": 1000000000.0,
            "minimum": 0.0,
            "title": "Upload Bytes",
            "type": "integer"
          },
          "window_seconds": {
            "default": 60,
            "maximum": 86400.0,
            "minimum": 1.0,
            "title": "Window Seconds",
            "type": "integer"
          }
        },
        "title": "PlanPolicy",
        "type": "object"
      },
      "PreviewInput": {
        "additionalProperties": false,
        "properties": {
          "expected_revision": {
            "maximum": 2000000000.0,
            "minimum": 0.0,
            "title": "Expected Revision",
            "type": "integer"
          },
          "policy": {
            "$ref": "#/components/schemas/PlanPolicy"
          }
        },
        "required": [
          "expected_revision",
          "policy"
        ],
        "title": "PreviewInput",
        "type": "object"
      },
      "PreviewView": {
        "additionalProperties": false,
        "properties": {
          "current_revision": {
            "title": "Current Revision",
            "type": "integer"
          },
          "diff": {
            "items": {
              "$ref": "#/components/schemas/DiffItem"
            },
            "title": "Diff",
            "type": "array"
          },
          "expected_revision": {
            "title": "Expected Revision",
            "type": "integer"
          },
          "generation_enabled": {
            "title": "Generation Enabled",
            "type": "boolean"
          },
          "impact": {
            "items": {
              "type": "string"
            },
            "title": "Impact",
            "type": "array"
          },
          "missing_required": {
            "items": {
              "type": "string"
            },
            "title": "Missing Required",
            "type": "array"
          }
        },
        "required": [
          "expected_revision",
          "current_revision",
          "diff",
          "generation_enabled",
          "missing_required",
          "impact"
        ],
        "title": "PreviewView",
        "type": "object"
      },
...
      "PublishInput": {
        "additionalProperties": false,
        "properties": {
          "expected_revision": {
            "maximum": 2000000000.0,
            "minimum": 0.0,
            "title": "Expected Revision",
            "type": "integer"
          },
          "operation_id": {
            "format": "uuid",
            "title": "Operation Id",
            "type": "string"
          },
          "policy": {
            "$ref": "#/components/schemas/PlanPolicy"
          },
          "reason": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "expected_revision",
          "policy",
          "operation_id",
          "reason"
        ],
        "title": "PublishInput",
        "type": "object"
      },
      "PublishReceipt": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "enum": [
              "publish",
              "rollback"
            ],
            "title": "Action",
            "type": "string"
          },
          "default_version": {
            "title": "Default Version",
            "type": "integer"
          },
          "operation_id": {
            "format": "uuid",
            "title": "Operation Id",
            "type": "string"
          },
          "policy_hash": {
            "title": "Policy Hash",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "revision_id": {
            "format": "uuid",
            "title": "Revision Id",
            "type": "string"
          }
        },
        "required": [
          "operation_id",
          "default_version",
          "revision_id",
          "revision",
          "policy_hash",
          "action"
        ],
        "title": "PublishReceipt",
        "type": "object"
      },
...
      "RollbackInput": {
        "additionalProperties": false,
        "properties": {
          "expected_revision": {
            "maximum": 2000000000.0,
            "minimum": 0.0,
            "title": "Expected Revision",
            "type": "integer"
          },
          "operation_id": {
            "format": "uuid",
            "title": "Operation Id",
            "type": "string"
          },
          "reason": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          },
          "target_revision": {
            "maximum": 2000000000.0,
            "minimum": 1.0,
            "title": "Target Revision",
            "type": "integer"
          }
        },
        "required": [
          "operation_id",
          "expected_revision",
          "target_revision",
          "reason"
        ],
        "title": "RollbackInput",
        "type": "object"
      },
...
      "SettingDescriptor": {
        "additionalProperties": false,
        "properties": {
          "apply_mode": {
            "const": "V",
            "default": "V",
            "title": "Apply Mode",
            "type": "string"
          },
          "default_semantics": {
            "title": "Default Semantics",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "required_when_generation_enabled": {
            "title": "Required When Generation Enabled",
            "type": "boolean"
          },
          "value_type": {
            "title": "Value Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "key",
          "value_type",
          "required_when_generation_enabled",
          "default_semantics"
        ],
        "title": "SettingDescriptor",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "IZO ASA Platform",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/settings/basic": {
      "get": {
        "operationId": "current_api_v1_admin_settings_basic_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicSettingsView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorView"
                }
              }
            },
            "description": "Safe settings error"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorView"
                }
              }
            },
            "description": "Safe settings error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorView"
                }
              }
            },
            "description": "Safe settings error"
          }
        },
        "summary": "Current",
        "tags": [
          "admin-settings"
        ]
      }
    },
    "/api/v1/admin/settings/basic/history": {
      "get": {
        "operationId": "history_api_v1_admin_settings_basic_history_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorView"
                }
              }
            },
            "description": "Safe settings error"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorView"
                }
              }
            },
            "description": "Safe settings error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorView"
                }
              }
            },
            "description": "Safe settings error"
          }
        },
        "summary": "History",
        "tags": [
          "admin-settings"
        ]
      }
    },
    "/api/v1/admin/settings/basic/preview": {
      "post": {
        "operationId": "preview_api_v1_admin_settings_basic_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorView"
                }
              }
            },
            "description": "Safe settings error"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorView"
                }
              }
            },
            "description": "Safe settings error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorView"
                }
              }
            },
            "description": "Safe settings error"
          }
        },
        "summary": "Preview",
        "tags": [
          "admin-settings"
        ]
      }
    },
    "/api/v1/admin/settings/basic/publish": {
      "post": {
        "operationId": "publish_api_v1_admin_settings_basic_publish_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishReceipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorView"
                }
              }
            },
            "description": "Safe settings error"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorView"
                }
              }
            },
            "description": "Safe settings error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorView"
                }
              }
            },
            "description": "Safe settings error"
          }
        },
        "summary": "Publish",
        "tags": [
          "admin-settings"
        ]
      }
    },
    "/api/v1/admin/settings/basic/rollback": {
      "post": {
        "operationId": "rollback_api_v1_admin_settings_basic_rollback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RollbackInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishReceipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorView"
                }
              }
            },
            "description": "Safe settings error"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorView"
                }
              }
            },
            "description": "Safe settings error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorView"
                }
              }
            },
            "description": "Safe settings error"
          }
        },
        "summary": "Rollback",
        "tags": [
          "admin-settings"
        ]
      }
    },
...
  }
}